Consistent toggling behaviour of language switching buttons
Since the language can be set both from the menu (via click or keyboard shortcut) as well as from the toolbar, all relevant buttons need to be synced. For example, when the menu button is clicked, the toolbar button has to be toggled from the event handling code. This prepares the GUI for implementing the translation mechanism in the future.