Added JSON data file download and open functionality
Added menu and toolbar buttons and keyboard shortcuts and the corresponding functionality to download COVID-19 timeseries JSON data and save it to a file as well as to load such files into the program. The implementation uses the wx.FileDialog for saving and opening the files, the urllib module for downloading the data and the json module for processing the JSON file. The download status is written to the statusText at the bottom of the window and cleared up after a certain delay using wx.CallLater.

This is a preparation for implementing data analysis and visualisation capabilities into the application.