Make country selection and plotting interactive, add more countries
The countries can be selected either via toggleable buttons in the toolbar or the menu. Both mechanism are kept in sync. There are quick-select buttons for selecting and deselecting all countries at once. When a country's selection state changes, the plots are redrawn, keeping the previously active tab in focus.

To make room for the preselected countries, the GUI was cleaned up by removing the nonsensical, mock-up buttons and menu entries. Likewise, the flag icons were removed from the language selection buttons by replacing them with a transparent 1x1 pixel bitmap.