Add static internationalisation / translation via wx
The internationalisation is based on the '_()' macro and uses 'pygettext' and 'msgfmt' to create the '.pot' and the '.mo' files. It is static, i.e. it is loaded during application start-up and cannot be changed during runtime.
Corresponding sections were added in the Sphinx documentation, both in the development instructions in 'README.rst' as well as the Notes section.

For convenience, the 'pygettext' and 'msgfmt' commands are included in a Makefile and can be run via 'make pot' and 'make mo'.