Add application settings via local config file
Switched to using Python's gettext module for the internationalisation because this allows selecting the language inside the main window's constructor after reading the configured language from the settings file.

The generate-po-en.py script creates an English .po file from the .pot file, but is not needed with the current language and settings architecture.