Properly show 'Strg' / 'Ctrl' in the German translation
In the menubar, use 'Ctrl' so that the keyboard shortcut is recognised.
In the helpString of the menubar, use 'Strg' because this is just text displayed in the statusbar.
In the shortHelp and longHelp of the toolbar, use 'Strg' because this is just text displayed in the tooltip and the statusbar, respectively.